Criando Log e retirando adm. do Reconhecimento
