+++ Login/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
   url(r'^obter_colaboradores/$', views.obter_colaboradores, name="obter_colaboradores"),
   url(r'^colaborador/$', views.inserir_colaboradores, name="colaborador"),
   url(r'^verificar_usuario/$', views.validar_usuario_logado, name="validar_usuario_logado"),
+  url(r'^administrador/$', views.switch_administrador, name="switch_administrador"),
 ]
